style: Code structure refactoring
