+++ app/src/main/java/kr/ac/deu/se/wisenote/ui/sign/SigninActivity.java
@@ -1,4 +1,4 @@
-package kr.ac.deu.se.wisenote;
+package kr.ac.deu.se.wisenote.ui.sign;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,10 +11,11 @@ import android.widget.Toast;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-
-import kr.ac.deu.se.wisenote.model.RetrofitClient;
-import kr.ac.deu.se.wisenote.model.ServiceApi;
-import kr.ac.deu.se.wisenote.model.signin.SigninResponse;
+import kr.ac.deu.se.wisenote.MainActivity;
+import kr.ac.deu.se.wisenote.R;
+import kr.ac.deu.se.wisenote.service.RetrofitClient;
+import kr.ac.deu.se.wisenote.service.ServiceApi;
+import kr.ac.deu.se.wisenote.vo.signin.SigninResponse;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
@@ -68,7 +69,7 @@ public class SigninActivity extends AppCompatActivity{
           SigninResponse result = response.body();
           result.setCode(response.code());
           Log.d("test","token:"+result.getAccess_token()+"code:"+result.getCode());
-          Intent intent = new Intent(SigninActivity.this,MainActivity.class);
+          Intent intent = new Intent(SigninActivity.this, MainActivity.class);
           startActivity(intent);
         }else if(response.code() == 401){
           Toast.makeText(SigninActivity.this,"ID 또는 PASSOWRD를 잘못입력하셨습니다.",Toast.LENGTH_SHORT).show();
